src/modules/item/rmbg_api/requirements.txt
fastapi
uvicorn
torch
transformers
numpy
Pillow
opencv-python-headless